requirements.txt
# Flask et extensions
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Login==0.6.3
Flask-WTF==1.1.1

# MySQL driver
PyMySQL==1.1.0

# Génération de PDF
WeasyPrint==59.0

# Gestion des mots de passe
bcrypt==4.0.1

# Pour la manipulation de fichiers et uploads
python-dotenv==1.0.0

# Librairie pour formulaires JSON / API
marshmallow==3.21.1

# Outils utilitaires
requests==2.32.1


Flask-Cors==1.10.1
python-slugify==8.0.1
PyJWT==2.8.0

